+++ src/main/java/com/raoulvdberge/refinedpipes/network/pipe/transport/ItemTransport.java
@@ -162,7 +162,7 @@ public class ItemTransport {
         }
 
         if (isLastPipe()) {
-            mt *= 0.50D;
+            mt *= 0.25D;
         }
 
         return (int) mt;